+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - added "units" (exact) and "dimensionality" (dimensionality compatible) checking options
   to `util.xr_check_coords` [[#442]](https://github.com/BAMWelDX/weldx/pull/442)
+- `Time` class that can be initialized from several other time types and provides time related utility functions 
+  [[#433]](https://github.com/BAMWelDX/weldx/pull/433)
 
 ### removed